pyproject.toml
[project]
name = "better-django-tables"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.3",
    "django-filter>=25.1",
    "django-tables2>=2.7.5",
]
